+++ CHANGELOG.md
@@ -3,7 +3,21 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased] 
+## [Unreleased]
+
+## [1.2.0] - 2020-10-15
+
+### Added
+
+- audit log for create, edit, delete 🕺
+    - can be turned off in global options `SchemaUI.init({ auditLog: false })`
+    - typings also added
+- improve tests
+
+### BugFixes
+
+- default boolean values were overridden
+- reduce the use of lodash 👌
 
 ## [1.1.1] - 2020-05-08 
 
@@ -45,7 +59,8 @@ Initial Version
 
 
 [unreleased]: https://github.com/slocking/schemaui/compare/master...dev
-[1.2.0]: https://github.com/slocking/schemaui/compare/v1.1.0...v1.1.1
+[1.2.0]: https://github.com/slocking/schemaui/compare/v1.1.1...v1.2.0
+[1.1.1]: https://github.com/slocking/schemaui/compare/v1.1.0...v1.1.1
 [1.1.0]: https://github.com/slocking/schemaui/compare/v1.0.5...v1.1.0
 [1.0.4]: https://github.com/slocking/schemaui/compare/v0.8.0...v1.0.4
 [0.8.0]: https://github.com/slocking/schemaui/compare/v0.8.0